requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==23.2  
gunicorn==22.0.0
idna==3.7
packaging==24.1
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
